+++ src/raw.rs
@@ -1,2 +1,0 @@
-//! Tools to interact with the raw MCP API.
-pub use rmcp::object as json_object;