+++ CHANGELOG.md
@@ -24,6 +24,8 @@ Headers:
 
 ## [Unreleased] - ReleaseDate
 
+## [0.0.5] - 2026-07-03
+
 ## Added
 
 - Low-level API that provides automatically generated functions from the MCP schema using `southesk-macros`.
@@ -86,7 +88,8 @@ Headers:
 First public release.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/thomasa88/southesk/compare/v0.0.4...HEAD
+[Unreleased]: https://github.com/thomasa88/southesk/compare/southesk-v0.0.5...HEAD
+[0.0.5]: https://github.com/thomasa88/southesk/compare/v0.0.4...southesk-v0.0.5
 [0.0.4]: https://github.com/thomasa88/southesk/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/thomasa88/southesk/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/thomasa88/southesk/compare/v0.0.1...v0.0.2